Move run and run_no_wait to traits
